backend/requirements.txt
flask==2.2.5
flask-cors==3.0.10
google-generativeai==0.3.0
python-dotenv==0.19.0
werkzeug==2.2.3 